Create class point
There are two classes called point2 and point3. The first one presents how to create a point with two fields - (x, y).  p1, p2, p3, p4 are objects of this class. The second class - point3 creates a point with three fields, using the first class as a type for the first two.